Tons of refactoring
retirejs.py now follows pep8 much better than it did before
and some of the logic flow is much more natural to follow.

main.go tries to follow go coding stylelines and some code is moved
in an attempt to make it much easier to understand and follow.

Semantics should barely be changed.